+++ RELEASE_NOTES.md
@@ -11,6 +11,14 @@ v1.0.0-alpha.x
   [#587](https://github.com/OpenAssetIO/OpenAssetIO/issues/587)
 
 
+### Improvements
+
+- Added `BatchElementError.ErrorCode.kEntityAccessError` to better
+  encode the scenario where the supplied context's access is the
+  specific cause of the error.
+  [#587](https://github.com/OpenAssetIO/OpenAssetIO/issues/587)
+
+
 v1.0.0-alpha.3
 --------------